% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_msavi}
\alias{calc_msavi}
\title{Calculate Modified Soil Adjusted Vegetation Index 2 (MSAVI)}
\usage{
calc_msavi(nir, red)
}
\arguments{
\item{nir}{Near-infrared band (\code{SpatRaster} or numeric).}

\item{red}{Red band (\code{SpatRaster} or numeric).}
}
\value{
A \code{SpatRaster} (named \code{"MSAVI"}) or numeric vector with index values.
}
\description{
Computes the Modified Soil Adjusted Vegetation Index 2, which eliminates the need
for finding the soil line factor \code{L} manually.
}
\details{
\deqn{MSAVI = \frac{2 \times NIR + 1 - \sqrt{(2 \times NIR + 1)^2 - 8 \times (NIR - RED)}}{2}}
}
\examples{
calc_msavi(nir = 0.7, red = 0.2)

}
\references{
Qi, J., et al. (1994). A modified soil adjusted vegetation index.
Remote Sensing of Environment, 48(2), 119-126.
}
