% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_msi}
\alias{calc_msi}
\title{Calculate Moisture Stress Index (MSI)}
\usage{
calc_msi(swir1, nir)
}
\arguments{
\item{swir1}{Short-wave infrared 1 band (\code{SpatRaster} or numeric).}

\item{nir}{Near-infrared band (\code{SpatRaster} or numeric).}
}
\value{
A \code{SpatRaster} (named \code{"MSI"}) or numeric vector with index values.
}
\description{
Computes the Moisture Stress Index from SWIR1 and NIR bands. Higher values indicate
higher plant water stress.
}
\details{
\deqn{MSI = \frac{SWIR1}{NIR}}
}
\examples{
calc_msi(swir1 = 0.4, nir = 0.6)

}
\references{
Rock, B. N., et al. (1986). Remote detection of forest damage. BioScience, 36(7), 439-445.
}
