% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_numeric_or_raster}
\alias{check_numeric_or_raster}
\title{Check that an input is either a SpatRaster or a numeric vector/array}
\usage{
check_numeric_or_raster(x, name = "band")
}
\arguments{
\item{x}{Input to check.}

\item{name}{Name of argument.}
}
\description{
Check that an input is either a SpatRaster or a numeric vector/array
}
\keyword{internal}
