% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_raster}
\alias{check_raster}
\title{Check that an object is a SpatRaster}
\usage{
check_raster(x, name = "image")
}
\arguments{
\item{x}{Object to check.}

\item{name}{Name of argument for error message.}
}
\description{
Check that an object is a SpatRaster
}
\keyword{internal}
