% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoIndexR-package.R
\docType{package}
\name{GeoIndexR-package}
\alias{GeoIndexR}
\alias{GeoIndexR-package}
\title{GeoIndexR: Computation of Spectral and Geospatial Indices from Multispectral Raster Data}
\description{
A unified, fast, and extensible framework for calculating spectral and geospatial indices from multispectral raster datasets (e.g., NDVI, NDWI, MNDWI, NDBI, SAVI, EVI, GNDVI, NDMI, BSI). Designed around 'terra' 'SpatRaster' objects, it supports multi-band rasters, automatic band resolution, sensor presets (Sentinel-2, Landsat-8/9), customizable index parameters, vectorized computations, and comprehensive validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sowsalim01/GeoIndexR}
  \item Report bugs at \url{https://github.com/sowsalim01/GeoIndexR/issues}
}

}
\author{
\strong{Maintainer}: Mamadou Sow \email{sowsalim01@gmail.com}

Authors:
\itemize{
  \item Mamadou Sow \email{sowsalim01@gmail.com}
}

}
\keyword{internal}
