% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_example_data}
\alias{get_example_data}
\title{Get Example Multispectral SpatRaster}
\usage{
get_example_data()
}
\value{
A \code{terra::SpatRaster} with 6 layers: \code{"blue"}, \code{"green"},
\code{"red"}, \code{"nir"}, \code{"swir1"}, \code{"swir2"}.
}
\description{
Returns the ready-to-use unpacked \code{terra::SpatRaster} representing the
bundled 6-band example dataset.
}
\examples{
img <- get_example_data()
print(img)
terra::nlyr(img)

}
