% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_numeric_or_raster}
\alias{check_numeric_or_raster}
\title{Check that an object is numeric or SpatRaster}
\usage{
check_numeric_or_raster(x, arg_name = "band")
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Name of argument.}
}
\value{
Invisible \code{TRUE} if valid; throws error otherwise.
}
\description{
Check that an object is numeric or SpatRaster
}
\keyword{internal}
