% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_raster}
\alias{check_raster}
\title{Check that an object is a SpatRaster}
\usage{
check_raster(x, arg_name = "image")
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Name of the argument in the calling function.}
}
\value{
Invisible \code{TRUE} if valid; throws an error otherwise.
}
\description{
Check that an object is a SpatRaster
}
\keyword{internal}
