% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_index_meta}
\alias{get_index_meta}
\title{Retrieve Metadata for a Single Spectral Index}
\usage{
get_index_meta(index)
}
\arguments{
\item{index}{Character string specifying index name (e.g., \code{"NDVI"}).}
}
\value{
A list containing detailed index metadata, parameter defaults, and formulas.
}
\description{
Retrieve Metadata for a Single Spectral Index
}
\keyword{internal}
